refactor: 25-move header_classes & createButtonWithSvg into separate files
